+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [Unreleased]
 ### Added
 - Support `docs`, `.github`, `.bitibucket` dirs as CODEOWNERS file locations
+- Support bitbucket config lines
 ## [v0.3.1](https://github.com/fan-tom/intellij-codeowners/tree/v0.3.1) (2021-11-25)
 ### Fixed
 - Bitbucket filetype detection